+++ src/main/java/org/morphix/reflection/Annotations.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -63,8 +63,8 @@ public class Annotations {
 			Map<String, Object> memberValues = Fields.IgnoreAccess.get(handler, FIELD_NAME_MEMBER_VALUES);
 			memberValues.put(attribute, value);
 		} catch (Exception e) {
-			throw new ReflectionException("Failed to override annotation: "
-					+ annotation.annotationType().getCanonicalName() + "." + attribute + "() value.", e);
+			throw new ReflectionException(e, "Failed to override annotation: {}.{}() value.",
+					annotation.annotationType().getCanonicalName(), attribute);
 		}
 	}