+++ src/main/java/org/morphix/reflection/GenericType.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -23,7 +23,16 @@ import java.util.StringJoiner;
 import org.morphix.lang.JavaObjects;
 
 /**
- * Implementation of {@link ParameterizedType}.
+ * Implementation of {@link ParameterizedType}. This class can be used to create parameterized types programmatically.
+ *
+ * <p>
+ * <b>Usage example:</b>
+ * <p>
+ * To create a parameterized type for <code>Map&lt;String, List&lt;Integer&gt;&gt;</code>:
+ *
+ * <pre>
+ * GenericType mapType = GenericType.of(Map.class, GenericType.Arguments.of(String.class, GenericType.of(List.class, Integer.class)));
+ * </pre>
  *
  * @author Radu Sebastian LAZIN
  */
@@ -45,6 +54,24 @@ public class GenericType implements ParameterizedType {
 	private final Type ownerType;
 
 	/**
+	 * Helper class to build type arguments arrays.
+	 *
+	 * @author Radu Sebastian LAZIN
+	 */
+	public interface Arguments {
+
+		/**
+		 * Builds a new array of {@link Type} from the given arguments.
+		 *
+		 * @param types types
+		 * @return array of types
+		 */
+		static Type[] of(final Type... types) {
+			return types;
+		}
+	}
+
+	/**
 	 * Private constructor with all arguments.
 	 *
 	 * @param rawType raw type
@@ -59,6 +86,8 @@ public class GenericType implements ParameterizedType {
 
 	/**
 	 * Builds a new {@link GenericType} object with the given arguments.
+	 * <p>
+	 * For the actual type arguments use {@link Arguments#of(Type...)} to build the array.
 	 *
 	 * @param rawType raw type
 	 * @param actualTypeArguments actual generic type arguments
@@ -82,7 +111,7 @@ public class GenericType implements ParameterizedType {
 	}
 
 	/**
-	 * Builds a new {@link GenericType} object with the given argument.
+	 * Builds a new {@link GenericType} object with the given argument. The owner type is set to null.
 	 *
 	 * @param rawType raw type
 	 * @param actualTypeArgument actual generic type argument
@@ -90,6 +119,19 @@ public class GenericType implements ParameterizedType {
 	 */
 	public static GenericType of(final Class<?> rawType, final Type actualTypeArgument) {
 		return of(rawType, actualTypeArgument, null);
+	}
+
+	/**
+	 * Builds a new {@link GenericType} object with the given arguments. The owner type is set to null.
+	 * <p>
+	 * For the actual type arguments use {@link Arguments#of(Type...)} to build the array.
+	 *
+	 * @param rawType raw type
+	 * @param actualTypeArguments actual generic type arguments
+	 * @return a new parameterized type object
+	 */
+	public static GenericType of(final Class<?> rawType, final Type[] actualTypeArguments) {
+		return of(rawType, actualTypeArguments, null);
 	}
 
 	/**
@@ -119,8 +161,7 @@ public class GenericType implements ParameterizedType {
 		if (genericArgumentType instanceof ParameterizedType parameterizedType) {
 			return of(parameterizedType);
 		}
-		throw new ReflectionException("Cannot build GenericType from " + genericArgumentType +
-				" because it is not a " + ParameterizedType.class);
+		throw new ReflectionException("Cannot build GenericType from {} because it is not a {}", genericArgumentType, ParameterizedType.class);
 	}
 
 	/**
@@ -133,17 +174,44 @@ public class GenericType implements ParameterizedType {
 	 * @return the generic parameter type for a given class
 	 */
 	public static <T extends Type> T getGenericParameterType(final Class<?> cls, final int index) {
-		if (index < 0) {
-			throw new ReflectionException("Generic parameter type index cannot be negative, received: " + index);
-		}
 		if (isNotGenericClass(cls)) {
-			throw new ReflectionException(cls.getCanonicalName() + " is not a generic class");
+			throw new ReflectionException("{} is not a generic class", cls.getCanonicalName());
 		}
 		ParameterizedType parameterizedType = (ParameterizedType) cls.getGenericSuperclass();
+		return getGenericParameterType(parameterizedType, cls.getCanonicalName(), index);
+	}
+
+	/**
+	 * Extract the generic parameter type for a given class.
+	 *
+	 * @param <T> generic parameter type
+	 *
+	 * @param parameterizedType parameterized type to extract from
+	 * @param index type parameter index
+	 * @return the generic parameter type for a given class
+	 */
+	public static <T extends Type> T getGenericParameterType(final ParameterizedType parameterizedType, final int index) {
+		return getGenericParameterType(parameterizedType, parameterizedType.getRawType().getTypeName(), index);
+	}
+
+	/**
+	 * Extract the generic parameter type for a given class.
+	 *
+	 * @param <T> generic parameter type
+	 *
+	 * @param parameterizedType parameterized type to extract from
+	 * @param typeName type name for error messages
+	 * @param index type parameter index
+	 * @return the generic parameter type for a given class
+	 */
+	public static <T extends Type> T getGenericParameterType(final ParameterizedType parameterizedType, final String typeName, final int index) {
+		if (index < 0) {
+			throw new ReflectionException("Generic parameter type index cannot be negative, received: {}", index);
+		}
 		Type[] actualTypeArguments = parameterizedType.getActualTypeArguments();
 		if (index >= actualTypeArguments.length) {
-			throw new ReflectionException("Cannot extract generic parameter type at index " + index + " from " + cls.getCanonicalName()
-					+ " because it has only " + actualTypeArguments.length + " generic parameter(s)");
+			throw new ReflectionException("Cannot extract generic parameter type at index {} from {} because it has only {} generic parameter(s)",
+					index, typeName, actualTypeArguments.length);
 		}
 		Type genericType = actualTypeArguments[index];
 		return JavaObjects.cast(genericType);
@@ -199,7 +267,7 @@ public class GenericType implements ParameterizedType {
 	}
 
 	/**
-	 * @see #getActualTypeArguments()
+	 * @see ParameterizedType#getActualTypeArguments()
 	 */
 	@Override
 	public Type[] getActualTypeArguments() {
@@ -207,7 +275,7 @@ public class GenericType implements ParameterizedType {
 	}
 
 	/**
-	 * @see #getRawType()
+	 * @see ParameterizedType#getRawType()
 	 */
 	@Override
 	public Type getRawType() {
@@ -215,7 +283,7 @@ public class GenericType implements ParameterizedType {
 	}
 
 	/**
-	 * @see #getOwnerType()
+	 * @see ParameterizedType#getOwnerType()
 	 */
 	@Override
 	public Type getOwnerType() {
@@ -223,7 +291,7 @@ public class GenericType implements ParameterizedType {
 	}
 
 	/**
-	 * @see #equals(Object)
+	 * @see Object#equals(Object)
 	 */
 	@Override
 	public boolean equals(final Object o) {
@@ -239,7 +307,7 @@ public class GenericType implements ParameterizedType {
 	}
 
 	/**
-	 * @see #hashCode()
+	 * @see Object#hashCode()
 	 */
 	@Override
 	public int hashCode() {
@@ -249,7 +317,7 @@ public class GenericType implements ParameterizedType {
 	}
 
 	/**
-	 * @see #toString()
+	 * @see Object#toString()
 	 */
 	@Override
 	public String toString() {